Merge cells around mesh points in each bin into pseudo-bulk gene expression matrix.

Merge cells around mesh points in each bin into pseudo-bulk gene expression matrix.

MergeCells(dge, gd, by = "mesh.points", cores = -1)

Arguments

dge

Matrix or sparse matrix for gene expression matrix rows are genes, columns are cells.

gd

A GridDensity object.

by

Merge raw counts by "bins" or "mesh.points". Default: "mesh.points"

cores

Number of threads, -1 means all available threads. Default: -1

Value

A gene expression matrix, rows are genes, columns are merged cells.