Visualizing the projected query cells on reference via a density plot.

Visualizing the projected query cells on reference via a density plot.

PlotProjection(
  seu.q,
  reference,
  ref.color.by,
  ref.colors = NULL,
  ref.alpha = 0.5,
  ref.size = 0.5,
  ref.emb.col = paste0("UMAP_", 1:2),
  query.reduction = "ref.umap",
  query.size = 1,
  query.alpha = 1,
  split.by = NULL,
  n.row = NULL,
  legend.ncol = 1
)

Arguments

seu.q

A query Seurat object containing the projected reference embeddings.

reference

Reference model.

ref.color.by

Name of column in reference data to be used for coloring points.

ref.colors

Vector of colors to be used for reference data points. Default: NULL

ref.alpha

Alpha value for reference data points. Default: 0.5

ref.size

Size of reference data points. Default: 0.5

ref.emb.col

The column names storing the reference embeddings in reference model.

query.reduction

Reduction name to use for query data. Default: 'ref.umap'

query.size

Size of query data points. Default: 1

query.alpha

Alpha value for query data points. Default: 1

split.by

Column name to split query data by. Default: NULL

n.row

Number of rows for facet_wrap when splitting by column. Default: NULL

legend.ncol

Number of columns for legend. Default: 1

Value

A ggplot object.